requirements.txt
# Core libraries for scraping and email validation
requests>=2.25.1
beautifulsoup4>=4.9.3
email-validator==2.1.0.post1
dnspython>=2.1.0
fake-useragent>=0.1.11
python-dotenv>=0.19.0
tweepy>=4.10.0
linkedin-api>=2.0.0

# Selenium for LinkedIn scraping
selenium>=4.1.0

# Optional: If running Selenium in Docker
webdriver-manager>=3.8.0  # To auto-manage ChromeDriver versions

# Additional libraries for compatibility in headless mode
pyvirtualdisplay==3.0  # For virtual display if needed